+++ metric/system/hwmon/hwmon.go
@@ -1,3 +1,20 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 package hwmon
 
 import (
@@ -5,128 +22,15 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"reflect"
 	"regexp"
 	"strconv"
 	"strings"
 
 	"github.com/elastic/elastic-agent-libs/opt"
 	"github.com/elastic/elastic-agent-system-metrics/metric/system/resolve"
-	"github.com/elastic/go-structform"
 )
 
-var baseDir = "/sys/class/hwmon"
-
-// Device represents a single sensor chip, usually exposed as /sys/class/hwmon/hwmon*
-type Device struct {
-	// Name is the specified hwmon label for the directory
-	Name string
-	// AbsPath is the absolute path to the monitoring directory, usually linked as /sys/class/hwmon/hwmon*
-	AbsPath string
-	//Sensors are the individual metrics connected to a device
-	Sensors []Sensor
-}
-
-// SensorType is for the string prefix of the sensor files
-type SensorType struct {
-	fileKey string
-	units   string
-}
-
-// TempSensor is the string prefix for a temp sensor
-var TempSensor = SensorType{fileKey: "temp", units: "celsius"}
-
-// VoltSensor is the prefix for voltage sensors
-var VoltSensor = SensorType{fileKey: "in", units: "millivolts"}
-
-// FanSensor is the prefix for fan sensors
-var FanSensor = SensorType{fileKey: "fan", units: "rpm"}
-
-// Sensor is used to track a single hwmon chip metric
-type Sensor struct {
-	DevType SensorType
-	// SensorNum is the numerical ID of the sensor, i.e temp7_*
-	SensorNum int64
-}
-
-// SensorMetrics reports the actual metrics in a sensors
-// This is meant to be generic for all possible sensor types,
-// hence the considerable heterogeneous fields
-type SensorMetrics struct {
-	//Generic Fields
-	Label string `struct:"label,omitempty"`
-
-	// This field gets inserted into the map, before individual values.
-	sensorType SensorType
-
-	Critical opt.Uint `struct:"critical,omitempty"`
-	Max      opt.Uint `struct:"max,omitempty"`
-	Lowest   opt.Uint `struct:"lowest,omitempty"`
-	Average  opt.Uint `struct:"average,omitempty"`
-
-	// The input value of the metric. The key is overriden and set to the value of sensorType by Fold()
-	Value opt.Uint `struct:"value,omitempty"`
-}
-
-// Fold implements the Folder interface for structform
-// This is entirely so we can carry around a relatively simple struct that transforms into the more heavily nested event that's the standard for beats events.
-func (sm *SensorMetrics) Fold(v structform.ExtVisitor) error {
-	val := reflect.ValueOf(sm).Elem()
-	types := reflect.TypeOf(sm).Elem()
-
-	v.OnObjectStart(val.NumField(), structform.AnyType)
-
-	for i := 0; i < val.NumField(); i++ {
-		if val.Field(i).CanInterface() {
-
-			// Fetch the struct tags
-			iface := val.Field(i).Interface()
-			skey, tagExists := types.Field(i).Tag.Lookup("struct")
-			if !tagExists {
-				skey = types.Field(i).Name
-			} else {
-				skey = strings.Split(skey, ",")[0]
-			}
-
-			if skey == "value" {
-				skey = sm.sensorType.fileKey
-			}
-
-			// Cast to an opt type, then create a nested dict
-			castUint, ok := iface.(opt.Uint)
-			if ok && castUint.Exists() {
-				err := v.OnKey(skey)
-				if err != nil {
-					return fmt.Errorf("error in OnKey for %s: %w", skey, err)
-				}
-				// This "inserts" the unit of the metric into the dict as an extra key
-				err = v.OnUint64Object(map[string]uint64{sm.sensorType.units: castUint.ValueOr(0)})
-				if err != nil {
-					return fmt.Errorf("error in OnKey for %s: %w", skey, err)
-				}
-
-			}
-		}
-
-	}
-	err := v.OnObjectFinished()
-	if err != nil {
-		return fmt.Errorf("error in OnObjectFinished for %s: %w", sm.Label, err)
-	}
-
-	return nil
-}
-
-// MonData is a simple wrapper type for the map returned by ReportSensors
-type MonData map[string]SensorMetrics
-
-// ErrNoMetric specifies that no metrics were found for a given Sensor.
-// This is meant to be a soft error if needed, as the slapdash nature of hwmon sysfs files
-// means that we can see *_label files with no corrisponding metrics.
-var ErrNoMetric = errors.New("No Metrics exist in this device")
-
 //ReportSensors returns the metrics from all the known sensors.
-//We would normally want a concrete data type here, but these metrics are so variable that we don't get much from it.
 func ReportSensors(dev Device) (MonData, error) {
 	metrics := MonData{}
 	for _, sensor := range dev.Sensors {
@@ -137,7 +41,6 @@ func ReportSensors(dev Device) (MonData, error) {
 			return nil, fmt.Errorf("error fetching sensor data for %s: %w", sensor.DevType, err)
 		}
 		// Create the device key from the label, a the values are considerably more intuative.
-		//sensorName := fmt.Sprintf("%s%d", sensor.DevType, sensor.SensorNum)
 		labelName := strings.ToLower(strings.Replace(data.Label, " ", "_", -1))
 		metrics[labelName] = data
 	}
@@ -161,16 +64,26 @@ func DetectHwmon(hostfs resolve.Resolver) ([]Device, error) {
 	}
 
 	sensorList := []Device{}
-	// the hwmon device directory is just a bunch of symlinks.
+
 	for _, path := range paths {
+
 		name := filepath.Join(fullPath, path.Name())
-		apath, err := os.Readlink(name)
+
+		apath := name
+		// This lstat call is largely to deal with the relative paths used by tests
+		stat, err := os.Lstat(name)
 		if err != nil {
-			return nil, fmt.Errorf("error reading path link %s: %w", name, err)
+			return nil, fmt.Errorf("error statting hwinfo path %s: %w", name, err)
 		}
 
-		if !filepath.IsAbs(apath) {
-			//These paths are usually relative, and filepath.Join will attempt to "clean" the realtive folders once we append an absolute base path
+		if stat.Mode()&os.ModeSymlink != 0 {
+			apath, err = os.Readlink(name)
+			if err != nil {
+				return nil, fmt.Errorf("error reading path link %s: %w", name, err)
+			}
+		}
+
+		if !filepath.IsAbs(apath) && stat.Mode()&os.ModeSymlink != 0 {
 			apath = filepath.Join(baseDir, apath)
 		}
 
@@ -180,7 +93,6 @@ func DetectHwmon(hostfs resolve.Resolver) ([]Device, error) {
 		}
 
 		namePath := filepath.Join(apath, "name")
-
 		sensorName, err := os.ReadFile(namePath)
 		if err != nil {
 			return nil, fmt.Errorf("error reading sensor name file %s: %w", namePath, err)
@@ -203,14 +115,9 @@ func (s Sensor) Fetch(path string) (SensorMetrics, error) {
 	labelName := s.getName("label")
 	label, err := stringStrip(labelName, path)
 
-	//Some directories will skip a label if there's only one sensor.
+	//Some sensors don't have a label, make our own
 	if os.IsNotExist(err) {
-		//try to fetch a name from the device symlink
-		cname, err := stringStrip("name", path)
-		if err != nil {
-			return SensorMetrics{}, fmt.Errorf("error reading name file: %w", err)
-		}
-		label = cname
+		label = fmt.Sprintf("%s_%d", s.DevType.fileKey, s.SensorNum)
 	} else if err != nil {
 		return SensorMetrics{}, fmt.Errorf("error fetching label for %s in %s: %w", labelName, path, err)
 	}
@@ -232,7 +139,7 @@ func (s Sensor) Fetch(path string) (SensorMetrics, error) {
 	}
 
 	// Other special metrics for some sensors
-	// We don't want to bulk fetch these with a glob or something, since the hwmon interface is so slapdash, we'll just end up picking up a bunch of garbage
+	// We don't want to bulk fetch these with a glob or something, we'll just end up picking up a bunch of garbage
 	critName := s.getName("crit")
 	critVal, _ := getAndDiv(critName, path, s.DevType)
 	sensorData.Critical = critVal
