+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/name
@@ -1,0 +1,1 @@
+coretemp
