+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp1_label
@@ -1,0 +1,1 @@
+Package id 0
