+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp2_label
@@ -1,0 +1,1 @@
+Core 0
