+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp4_input
@@ -1,0 +1,1 @@
+50000
