+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp4_max
@@ -1,0 +1,1 @@
+81000
