+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp5_crit_alarm
@@ -1,0 +1,1 @@
+0
