+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp5_input
@@ -1,0 +1,1 @@
+53000
