+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp6_input
@@ -1,0 +1,1 @@
+52000
