+++ metric/system/hwmon/testdata/PoweredgeR720/sys/class/hwmon/hwmon0/temp7_crit
@@ -1,0 +1,1 @@
+91000
