+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon3/in0_input
@@ -1,0 +1,1 @@
+11943
