+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon3/name
@@ -1,0 +1,1 @@
+BAT1
