+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/fan1_input
@@ -1,0 +1,1 @@
+2200
