+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/name
@@ -1,0 +1,1 @@
+thinkpad
