+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/pwm1_enable
@@ -1,0 +1,1 @@
+2
