+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/temp1_label
@@ -1,0 +1,1 @@
+CPU
