+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/temp2_label
@@ -1,0 +1,1 @@
+GPU
