+++ metric/system/hwmon/testdata/ThinkpadX250/sys/class/hwmon/hwmon4/temp6_input
@@ -1,0 +1,1 @@
+76000
